refactor(cli): group helper workspace-info commands into workspaceinfo sub-package
