+++ cmd/internal/helper/helper.go
@@ -7,6 +7,7 @@ import (
 	"github.com/devsy-org/devsy/cmd/internal/helper/json"
 	"github.com/devsy-org/devsy/cmd/internal/helper/ssh"
 	"github.com/devsy-org/devsy/cmd/internal/helper/strings"
+	"github.com/devsy-org/devsy/cmd/internal/helper/workspaceinfo"
 	"github.com/spf13/cobra"
 )
 
@@ -25,9 +26,9 @@ func NewHelperCmd(globalFlags *flags.GlobalFlags) *cobra.Command {
 	helperCmd.AddCommand(json.NewJSONCmd(globalFlags))
 	helperCmd.AddCommand(strings.NewStringsCmd(globalFlags))
 	helperCmd.AddCommand(ssh.NewSSHServerCmd(globalFlags))
-	helperCmd.AddCommand(NewGetWorkspaceNameCmd(globalFlags))
-	helperCmd.AddCommand(NewGetWorkspaceUIDCmd(globalFlags))
-	helperCmd.AddCommand(NewGetWorkspaceConfigCommand(globalFlags))
+	helperCmd.AddCommand(workspaceinfo.NewGetWorkspaceNameCmd(globalFlags))
+	helperCmd.AddCommand(workspaceinfo.NewGetWorkspaceUIDCmd(globalFlags))
+	helperCmd.AddCommand(workspaceinfo.NewGetWorkspaceConfigCommand(globalFlags))
 	helperCmd.AddCommand(NewGetProviderNameCmd(globalFlags))
 	helperCmd.AddCommand(NewCheckProviderUpdateCmd(globalFlags))
 	helperCmd.AddCommand(ssh.NewSSHClientCmd())
