+++ cmd/internal/helper/workspaceinfo/get_workspace_config.go
@@ -1,4 +1,4 @@
-package helper
+package workspaceinfo
 
 import (
 	"context"
