+++ cmd/internal/helper/workspaceinfo/get_workspace_uid.go
@@ -1,4 +1,4 @@
-package helper
+package workspaceinfo
 
 import (
 	"context"
